% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{Greatest common factor.}
\usage{
euclidian(x, y)
}
\arguments{
\item{...}{Numeric,integers}
}
\value{
if all the input values are integers or numeric it returns the greatest common divisor if the input is something else it returns error
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm} for more details.
}
\description{
\code{euclidian function} returns the greatest common factor present in its arguments.
euclidian function takes two intergers or maybe numbers and it returns the greatest common factor of the
}
\examples{
euclidian(10,35)
}

