\name{oureuc-package}
\alias{oureuc-package}
\alias{oureuc}
\docType{package}
\title{
\packageTitle{oureuc}
}
\description{
\packageDescription{oureuc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{oureuc}
\packageIndices{oureuc}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{oureuc}

Maintainer: \packageMaintainer{oureuc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
